Implement CoinListViewModel and CoinListScreen
This commit introduces the `CoinListViewModel` for managing the coin list state and the `CoinListScreen` for displaying it. It also sets up Koin for dependency injection and configures networking with Ktor. The `MainActivity` is updated to use the `CoinListScreen`.
Additionally it introduces `CryptoTrackerApp` class for Koin initialisation, adds `INTERNET` permission, and removes the previous Greeting preview.